requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dill==0.3.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==3.13.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.2.0
promise==2.3
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-metadata==1.0.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
